added absolute length switch and selector of range(not functional)
