+++ main.py
@@ -1,5 +1,6 @@
 from flask import Flask, render_template, redirect, url_for, flash, request, session, jsonify
-from dataframe import GermlineAnalyzer, files_to_dictionary, read_mitotic_file_into_average, extract_length
+from dataframe import GermlineAnalyzer, files_to_dictionary, read_mitotic_file_into_average, \
+    extract_length, calculate_average_length, determine_same_length_units
 from grapher import plotGermline, convert_plot_to_png, encode_png_to_base64
 import pandas as pd
 import os
@@ -146,30 +147,8 @@ def plot(strains):
 
     #extract lenght info
     final_length_list = extract_length(dataframes)
-
-    def determine_same_length_units(l_list_list):
-        units = set()
-        for strain in l_list_list:
-            for length in strain:
-                units.add(length[1])
-        return len(units) == 1
-
-    def calculate_average_length(l_list_list, s_name_list):
-        average_list = []
-        for i , strain in enumerate(l_list_list):
-            strain_average = 0
-            units = set()
-            for length in strain:
-                strain_average += length[0]
-                units.add(length[1])
-            if len(units) == 1:
-                units = list(units)
-                average_list.append([s_name_list[i], round(strain_average/len(strain), 1), units[0]])
-            else:
-                average_list.append([None,"Different units in files"])
-        return average_list
-
     average_length = calculate_average_length(final_length_list, strain_name_list)
+    can_absolute_length = determine_same_length_units(final_length_list)
     print("average", average_length)
     print("final length list", final_length_list)
 
@@ -215,12 +194,14 @@ def plot(strains):
         b64 = encode_png_to_base64(png)
 
         return render_template('plot.html', strains=strains, image=b64, files_list_list=files_list_list,
-                               strain_name_list=strain_name_list, lengths_list_list=final_length_list, npoints=int(points), range_fold=per_fld,
+                               strain_name_list=strain_name_list, lengths_list_list=final_length_list,
+                               average_length=average_length, can_absolute_length=can_absolute_length,
+                               npoints=int(points), range_fold=per_fld,
                                range_fold_1=int(range_start), range_fold_2=int(range_end), std=std, fld=fld, dpi=dpi,
                                mitotic=mitotic_files_loaded, mitotic_switched_on=mitotic_switched_on)
 
     return render_template('plot.html', strains=strains, files_list_list=files_list_list,strain_name_list=strain_name_list,
-                           lengths_list_list=final_length_list, average_length=average_length,
+                           lengths_list_list=final_length_list, average_length=average_length, can_absolute_length=can_absolute_length,
                            npoints=50, range_fold="0-4", range_fold_1=0, range_fold_2=4, std=False, fld=False, dpi=100,
                            mitotic=mitotic_files_loaded, mitotic_switched_on=True)
