+++ templates/footer.html
@@ -114,6 +114,9 @@
     function updateTextInput(val) {
           document.getElementById('textInput').value=val;
         }
+    function updateTextInput1(val) {
+        document.getElementById('textInput1').value=val;
+    }
     function updateTextInput2(val) {
           document.getElementById('textInput22').value=val;
         }
